+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.4.0 (in development)
 - Support python-dotenv v1.0
 - Use [`ghtoken`](https://github.com/jwodder/ghtoken) for looking up GitHub
   tokens
+- Ensure that the state file is always read & written using UTF-8 encoding
 
 v0.3.0 (2023-02-09)
 -------------------